% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_des}
\alias{add_des}
\title{Add descriptive table to the object.}
\usage{
add_des(x, file)
}
\arguments{
\item{x}{An object of the class mina with @tab defined.}

\item{file}{The descriptive table.}
}
\value{
x An object of the class mina with @des_tab added.
}
\description{
This includes functions for adding descriptive table (e.g. design file)
to mina object for later analysis. The samples in first column "Sample_ID"
should be the same as the samples in the quantitative table.
}
\examples{
add_des(x, "design.txt")

}
