% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_norm}
\alias{add_norm}
\title{Add normalized quantitative table to object mina}
\usage{
add_norm(mina, norm)
}
\arguments{
\item{mina}{An object of the class mina.}

\item{norm}{The normalized quantitative table or the file contains the table.}
}
\value{
mina An object of the class mina with normalized quantitative table
added.
}
\description{
The normalized quantitative table should contains the exactly same samples
as quantitative table and descriptive table in the same mina object.
}
\examples{
add_norm(mina, norm)
add_norm(mina, "otu_table_norm.txt")
}
