% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_tab}
\alias{add_tab}
\title{Add tables to the object}
\usage{
add_tab(mina, file)
}
\arguments{
\item{mina}{An object of the class mina.}

\item{file}{The quantitative table.}
}
\value{
mina An object of the class mina with quantitative table added.
}
\description{
This includes functions for adding quantitative table (e.g. OTU table)
to mina object for later analysis.
}
\examples{
add_tab(mina, "otu_table.txt")

}
