% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addTab}
\alias{addTab}
\title{Add tables to the object}
\usage{
addTab(mina, file)
}
\arguments{
\item{mina}{An object of the class mina.}

\item{file}{The quantitative table.}
}
\value{
mina with quantitative table.
}
\description{
This includes functions for adding quantitative table (e.g. OTU table)
to mina object for later analysis.
}
\examples{
addTab(mina, "otu_table.txt")

}
