% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\docType{methods}
\name{adj,matrix,ANY-method}
\alias{adj,matrix,ANY-method}
\title{Calculate the adjacacency matrix of @norm by correlation with matrix as
input.}
\usage{
\S4method{adj}{matrix,ANY}(x, method, threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An matrix for correlation / adjacency matrix calculation.}

\item{method}{The correlation coefficient used for adjacacency matrix.}

\item{threads}{(optional) The number of threads used for parallel running, 80
by default.}

\item{nblocks}{(optional) The number of row / column for splitted sub-matrix,
400 by default.}
}
\value{
y The adjacacency matrix.
}
\description{
Calculate the adjacacency matrix of @norm by correlation with matrix as
input.
}
\examples{
y <- adj(x, method = "pearson")
y <- adj(x, method = "spearman")
y <- adj(x, method = "sparcc", threads = 80, nblocks = 400)
}
