% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\docType{methods}
\name{adj,matrix,ANY,ANY-method}
\alias{adj,matrix,ANY,ANY-method}
\alias{adj,matrix,character,ANY-method}
\title{Calculate the adjacency matrix of @norm by correlation with matrix as input.}
\usage{
\S4method{adj}{matrix,ANY,ANY}(x, method, threads = 80, nblocks = 400)

\S4method{adj}{matrix,character,ANY}(x, method, threads = 80,
  nblocks = 400)
}
\arguments{
\item{x}{An matrix for correlation/adjacency matrix calculation.}

\item{method}{The correlation coefficient used for adjacency matrix.}

\item{threads}{The number of threads used for parallel running, 80 by
default.}

\item{nblocks}{The number of row/column for splitting sub-matrix, 400 by
default.}
}
\value{
y The adjacency matrix.
}
\description{
Calculate the adjacency matrix of @norm by correlation with matrix as input.
}
\examples{
\dontrun{
data(maize_asv)
asv <- maize_asv[1 : 500, 1 : 300]
asv_adj <- adj(asv, method = "sparcc", threads = 2, nblocks = 40)
asv_norm <- norm_tab(asv, method = "raref", depth = 100)
asv_adj <- adj(asv_norm, method = "pearson")
asv_adj <- adj(asv_norm, method = "spearman")
}
}
