% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\docType{methods}
\name{adj,mina,ANY-method}
\alias{adj,mina,ANY-method}
\title{Calculate the adjacacency matrix of @norm by correlation with `mina` class
object as input.}
\usage{
\S4method{adj}{mina,ANY}(x, method, threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An object of the class mina with @norm defined.}

\item{method}{The correlation coeffient used for adjacacency matrix.}

\item{threads}{(optional) The number of threads used for parallel running, 80
by default.}

\item{nblocks}{(optional) The number of row / column for splitted sub-matrix,
400 by default.}
}
\value{
x The same `mina` object with @adj added.
}
\description{
Calculate the adjacacency matrix of @norm by correlation with `mina` class
object as input.
}
\examples{
x <- adj(x, method = "pearson")
x <- adj(x, method = "spearman", thread = 80, nblocks = 400)
x <- adj(x, method = "sparcc", thread = 40, nblocks = 200)
}
