% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\docType{methods}
\name{adj-mina}
\alias{adj-mina}
\alias{adj,mina,character-method}
\title{Calculate the adjacacency matrix of @norm by correlation with `mina` class
object as input.}
\usage{
\S4method{adj}{mina,ANY}(x, method, threads = 80, nblocks = 400)

\S4method{adj}{mina,character}(x, method, threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An object of the class mina with @norm defined.}

\item{method}{The correlation coeffient used for adjacacency matrix.}

\item{threads}{The number of threads used for parallel running, 80 by
default.}

\item{nblocks}{The number of row / column for splitted sub-matrix, 400 by
default.}
}
\value{
x The same `mina` object with @adj added.
}
\description{
Calculate the adjacacency matrix of @norm by correlation with `mina` class
object as input.
}
\examples{
\dontrun{
data(maize)
maize <- norm_tab(maize, method = "raref")
maize <- adj(maize, method = "pearson")
}
}
