% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{.adj<-}
\alias{.adj<-}
\alias{.adj<-,mina-method}
\alias{.adj}
\alias{adj_sig<-}
\alias{adj_sig<-,mina-method}
\alias{adj_sig}
\title{Setter for the slot `adj` and `adj_sig`, the adjacency matrix of
`norm` and corresponding significant value matrix with `sig` is `TRUE`.}
\usage{
.adj(x) <- value

\S4method{.adj}{mina}(x) <- value

.adj(x)

adj_sig(x) <- value

\S4method{adj_sig}{mina}(x) <- value

adj_sig(x)
}
\arguments{
\item{x}{The `mina` object.}

\item{value}{The value to set for the slot of the `mina` object `x`.}
}
\value{
The `adj` slot of the `mina` object.

The slot `adj_sig` of the object.
}
\description{
Setter for the slot `adj` and `adj_sig`, the adjacency matrix of
`norm` and corresponding significant value matrix with `sig` is `TRUE`.

Get the slot `adj`.

Get the slot `adj_sig`.
}
\keyword{internal}
