% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_pm.R
\docType{methods}
\name{bs_pm,mina,ANY,numeric,numeric,logical,logical,numeric,numeric-method}
\alias{bs_pm,mina,ANY,numeric,numeric,logical,logical,numeric,numeric-method}
\title{Inferring the network of different group of samples and test significance by
permutation.}
\usage{

  \S4method{bs_pm}{mina,ANY,numeric,numeric,logical,logical,numeric,numeric}(x,
  group, g_size = 88, s_size = 30, rm = TRUE, sig = TRUE, bs = 6,
  pm = 6)
}
\arguments{
\item{x}{An object of class `mina` with @norm and @des defined.}

\item{group}{The column name of descriptive file @des for comparison.}

\item{g_size}{The cutoff of group size used for filtering, default 88.}

\item{s_size}{The number of samples used for network inference during
bootstrap and permutation (when sig == TRUE), it should be smaller than
g_size / 2 to make sure the randomness; default 30.}

\item{rm}{Filtering the components present in less than 20% of the samples
from compared groups, default TRUE.}

\item{sig}{Whether to test the significance, skip the permutation when sig ==
FALSE, default TRUE.}

\item{bs}{The times for bootstrap network inference, default 6}

\item{pm}{The times for permuatated samples network inference, default 6.}
}
\value{
x The same object with @multi and @perm defined.
}
\description{
Inferring the network of different group of samples and test significance by
permutation.
}
\examples{
x <- bs_pm(x, group = "Compartment")
x <- bs_pm(x, group = "Compartment", g_size = 100, s_size = 50, rm = TRUE,
sig = TRUE, bs = 6, pm = 6)
}
