% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{multi<-}
\alias{multi<-}
\alias{multi}
\alias{multi<-,mina-method}
\alias{multi,mina-method}
\alias{perm<-}
\alias{perm}
\alias{perm<-,mina-method}
\alias{perm,mina-method}
\title{Setter and getter for the slot `multi` and `perm`.}
\usage{
multi(x) <- value

multi(x)

\S4method{multi}{mina}(x) <- value

\S4method{multi}{mina}(x)

perm(x) <- value

perm(x)

\S4method{perm}{mina}(x) <- value

\S4method{perm}{mina}(x)
}
\arguments{
\item{x}{The `mina` object.}

\item{value}{The value to set for the slot of the `mina` object `x`.}
}
\value{
The `multi` or/and `perm` slot of the `mina` object.
}
\description{
Setter and getter for the slot `multi` and `perm`.
}
\keyword{internal}
