% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mina}
\alias{check_mina}
\title{Check the Tab and desTab file, used in setClass. Return TRUE if they are NULL or both quantitative and 
and descriptive files of same samples are included (i.e. the object is valid).}
\usage{
check_mina(object)
}
\arguments{
\item{object}{a mina object}
}
\value{
TRUE if the object is valid or empty.
}
\description{
Check the Tab and desTab file, used in setClass. Return TRUE if they are NULL or both quantitative and 
and descriptive files of same samples are included (i.e. the object is valid).
}
