% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mina_de}
\alias{check_mina_de}
\title{Check the object and return TRUE if the object includes descriptive table
contains the same samples as quantitative table.}
\usage{
check_mina_de(x)
}
\arguments{
\item{x}{An object of class mina with @tab and @des defined.}
}
\value{
TRUE if the object contains non-empty descriptive table and has the
same samples as quantitative table.
}
\description{
Check the object and return TRUE if the object includes descriptive table
contains the same samples as quantitative table.
}
\examples{
data(maize)
check_mina_de(maize)
}
