% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mina_qu}
\alias{check_mina_qu}
\title{Check the object and return TRUE if the object includes quantitative table.}
\usage{
check_mina_qu(x)
}
\arguments{
\item{x}{An object of class mina with @tab defined.}
}
\value{
TRUE if the object contains quantitative table and is not empty.
}
\description{
Check the object and return TRUE if the object includes quantitative table.
}
\examples{
x <- new("mina", tab = maize_asv)
check_mina_qu(x)
}
\keyword{internal}
