% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{cls}
\alias{cls}
\title{Get the slot `cls`.}
\usage{
cls(x)
}
\arguments{
\item{x}{The `mina` object.}
}
\value{
The `cls` slot of the object.
}
\description{
Get the slot `cls`.
}
\keyword{internal}
