% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{cls<-}
\alias{cls<-}
\alias{cls<-,mina-method}
\title{Setter for the slot `cls`.}
\usage{
cls(x) <- value

\S4method{cls}{mina}(x) <- value
}
\arguments{
\item{x}{The `mina` object.}

\item{value}{The value to set for the slot of the `mina` object `x`.}
}
\description{
Setter for the slot `cls`.
}
\keyword{inter}
\keyword{internal}
