% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{cls_tab}
\alias{cls_tab}
\title{Get the slot `cls_tab`.}
\usage{
cls_tab(x)
}
\arguments{
\item{x}{The `mina` object.}
}
\value{
The `cls_tab` slot of the object.
}
\description{
Get the slot `cls_tab`.
}
\examples{
cls_tab(maize)
}
