% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{cls_tab<-}
\alias{cls_tab<-}
\alias{cls_tab<-,mina-method}
\title{Setter for the slot `cls_tab`.}
\usage{
cls_tab(x) <- value

\S4method{cls_tab}{mina}(x) <- value
}
\arguments{
\item{x}{The `mina` object.}

\item{value}{The value to set for the slot of the `mina` object `x`.}
}
\description{
Setter for the slot `cls_tab`.
}
\keyword{internal}
