% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_dis.R
\docType{methods}
\name{com_dis,matrix,ANY-method}
\alias{com_dis,matrix,ANY-method}
\title{Calculate the community dissimilarity / distance matrix of the input matrix.}
\usage{
\S4method{com_dis}{matrix,ANY}(x, method, threads = 80, nblocks = 400)
}
\arguments{
\item{x}{A matrix of the quantitative table.}

\item{method}{The dissimilarity / distance method used.}

\item{threads}{(optional) The number of threads used for parallel running.}

\item{nblocks}{(optional) The number of row / column for splitted sub-matrix.}
}
\value{
y The dissimilarity / distance matrix.
}
\description{
Calculate the community dissimilarity / distance matrix of the input matrix.
}
\examples{
y <- com_dis(x, method = "bray")
y <- com_dis(x, method = "tina", threads = 40, nblocks = 200)
}
