% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_dis.R
\docType{methods}
\name{com_dis,mina,ANY-method}
\alias{com_dis,mina,ANY-method}
\title{Calculate the community dissimilarity / distance matrix of @norm with `mina`
class object as input.}
\usage{
\S4method{com_dis}{mina,ANY}(x, method, threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An object of the class `mina` with @norm defined.}

\item{method}{The dissimilarity / distance method used.}

\item{threads}{(optional, only needed when method == "tina") The number of
threads used for parallel running.}

\item{nblocks}{(optional, only needed when method == "tina") The number of
row / column for splitted sub-matrix.'}
}
\value{
x The same `mina` object with @dis added.
}
\description{
Calculate the community dissimilarity / distance matrix of @norm with `mina`
class object as input.
}
\examples{
x <- com_dis(x, method = "bray")
x <- com_dis(x, method = "tina", threads = 40, nblocks = 200)
}
