% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{com_dis}
\alias{com_dis}
\title{Calculate the community dissimilarity / distance matrix of @norm.}
\usage{
com_dis(x, method = "bray", threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An object of the class mina with @norm defined or any quantitative
matrix.}

\item{method}{The dissimilarity / distance method used, default `bray`.}

\item{threads}{The number of threads used for parallel running, needed for
method `tina`.}

\item{nblocks}{The number of row / column for splitted sub-matrix, needed for
method `tina`.}
}
\description{
Calculate the community dissimilarity / distance matrix of @norm.
}
\examples{
data(maize)
maize@tab <- maize@tab[1 : 500, 1 : 200]
maize <- norm_tab(maize, method = "raref", depth = 200)
maize <- fit_tabs(maize)
maize <- com_dis(maize, method = "bray")
maize <- com_dis(maize, method = "tina", threads = 2, nblocks = 40)
}
