% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_dis.R
\docType{data}
\name{com_dis_list}
\alias{com_dis_list}
\title{List of dissimilarity / distance supported in \code{\link[mina]{com_dis}}}
\format{A list of character vectors.

\describe{
  \item{\code{tina}}{TINA from Schmidt_et_al_2016}

  \item{\code{manhattan}}{\code{\link[vegan]{vegdist}}}
  \item{\code{euclidean}}{\code{\link[vegan]{vegdist}}}
  \item{\code{canberra}}{\code{\link[vegan]{vegdist}}}
  \item{\code{bray}}{\code{\link[vegan]{vegdist}}}
  \item{\code{kulczynski}}{\code{\link[vegan]{vegdist}}}
  \item{\code{jaccard}}{\code{\link[vegan]{vegdist}}}
  \item{\code{gower}}{\code{\link[vegan]{vegdist}}}
  \item{\code{altGower}}{\code{\link[vegan]{vegdist}}}
  \item{\code{morisita}}{\code{\link[vegan]{vegdist}}}
  \item{\code{horn}}{\code{\link[vegan]{vegdist}}}
  \item{\code{mountford}}{\code{\link[vegan]{vegdist}}}
  \item{\code{raup}}{\code{\link[vegan]{vegdist}}}
  \item{\code{binomial}}{\code{\link[vegan]{vegdist}}}
  \item{\code{chao}}{\code{\link[vegan]{vegdist}}}
  \item{\code{cao}}{\code{\link[vegan]{vegdist}}}
}}
\usage{
com_dis_list
}
\description{
Dissimilarity / distance should be specified by exact string match.
}
\examples{
com_dis_list
}
\keyword{datasets}
