% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_plot.R
\docType{methods}
\name{com_plot,mina,ANY,ANY-method}
\alias{com_plot,mina,ANY,ANY-method}
\title{Visulization of components distance / dissimilarity in k dimension.}
\usage{
\S4method{com_plot}{mina,ANY,ANY}(x, match, color, shape)
}
\arguments{
\item{x}{An object of `mina` with list @dmr defined.}

\item{match}{The column name of the components IDs in @des with exactly the
same as rownames in x.}

\item{color}{The column name in @des to be used for different color groups.}

\item{(optional)}{shape The column name in @des to be used for different
shape groups.}
}
\value{
p The plotted figure.
}
\description{
Visulization of components distance / dissimilarity in k dimension.
}
\examples{
p <- com_plot(x, match = "Sample_ID", color = "Compartment",
              shape = "Soil")
}
