% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{com_plot}
\alias{com_plot}
\title{Visulization of components distance / dissimilarity in k dimension.}
\usage{
com_plot(x, match, color, shape = NULL)
}
\arguments{
\item{x}{An object of class `mina` with @dmr and @des defined.}

\item{match}{The column name of the components IDs in @des which exactly the
same indicated in @dmr.}

\item{color}{The column name in @des to be used for different color groups.}

\item{(optional)}{shape The column name in @des to be used for different
shape groups.}
}
\description{
Visulization of components distance / dissimilarity in k dimension.
}
\examples{
p <- com_plot(x, match = "Sample_ID", color = "Compartment", shape = "Soil")
}
