% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\docType{methods}
\name{com_r2,mina,ANY-method}
\alias{com_r2,mina,ANY-method}
\title{Function for unexplained variance ratio calculation indicated in
Anderson, M.J. 2001. A new method for non-parametric multivariate analysis of
variance. Austral Ecology, 26: 32--46.}
\usage{
\S4method{com_r2}{mina,ANY}(x, group)
}
\arguments{
\item{x}{An mina object with @dis and @des_tab defined.}

\item{group}{The name(s) of column(s) defined as experimental setup group(s).}
}
\value{
r2 The variance ratio cannot be explained by given groups.
}
\description{
Function for unexplained variance ratio calculation indicated in
Anderson, M.J. 2001. A new method for non-parametric multivariate analysis of
variance. Austral Ecology, 26: 32--46.
}
\examples{
x <- com_r2(x, group = "Soil")

m <- new("mina", dis = distance_matrix, des_tab = design_data_frame)
m <- com_r2(m)

}
