% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\name{cor_cpp}
\alias{cor_cpp}
\title{Rcpp version function for `pearson` and `spearman` correlation calculation,
modified from https://systematicinvestor.github.io/Correlation-Rcpp
RcppParallel is used for parallel running.}
\usage{
cor_cpp(x, method = "pearson")
}
\arguments{
\item{x}{An matrix for correlation calculation.}

\item{method}{The correlation coeffient used for adjacacency matrix.}
}
\value{
y The adjacacency matrix.
}
\description{
Rcpp version function for `pearson` and `spearman` correlation calculation,
modified from https://systematicinvestor.github.io/Correlation-Rcpp
RcppParallel is used for parallel running.
}
\examples{
y <- cpp_cor(x, method = "pearson")
}
