% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\name{cor_par}
\alias{cor_par}
\title{Function for `pearson` and `spearman` correlation calculation. Modified from
https://github.com/defleury/Schmidt_et_al_2016_community_similarity/blob/
master/functions.community_similarity.R}
\usage{
cor_par(x, method = "pearson", threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An matrix for correlation calculation.}

\item{method}{The correlation coeffient used for adjacacency matrix.}

\item{threads}{(optional) The number of threads used for parallel running, 80
by default.}

\item{nblocks}{(optional) The number of row / column for splitted sub-matrix,
400 by default.}
}
\value{
y The adjacacency matrix.
}
\description{
Function for `pearson` and `spearman` correlation calculation. Modified from
https://github.com/defleury/Schmidt_et_al_2016_community_similarity/blob/
master/functions.community_similarity.R
}
\examples{
y <- cor_par(x, method = "pearson", threads = 80, nblocks = 400)
}
