% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{cp_cor}
\alias{cp_cor}
\title{Calculate the Person / Spearman correlation of @norm.}
\usage{
cp_cor(x)
}
\arguments{
\item{x}{An matrix for pearson correlation calculation.}
}
\value{
y The adjacacency matrix.
}
\description{
Calculate the Person / Spearman correlation of @norm.
}
\examples{
y <- cp_cor(x)
}
\keyword{internal}
