% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-hmp}
\alias{data-hmp}
\alias{hmp}
\title{Internal testing data of HMP project, including quantitative table (hmp_otu)
and descriptive table (hmp_des) for testing.}
\description{
Internal testing data of HMP project, including quantitative table (hmp_otu)
and descriptive table (hmp_des) for testing.
}
\examples{
data(hmp)
}
\keyword{data}
