% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data-maize}
\alias{data-maize}
\alias{maize}
\title{Internal testing data of maize project, vegetative stage samples only,
including quantitative table (maize_asv.rds) and descriptive table
(maize_des.txt) for testing.}
\description{
Internal testing data of maize project, vegetative stage samples only,
including quantitative table (maize_asv.rds) and descriptive table
(maize_des.txt) for testing.
}
\examples{
data(maize)
}
\keyword{data}
