% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{des<-}
\alias{des<-}
\alias{des<-,mina-method}
\alias{des}
\alias{des,mina-method}
\title{Setter and getter for the slot `des`, which is the description and meta data
of rows in `tab`.}
\usage{
des(x) <- value

\S4method{des}{mina}(x) <- value

des(x)

\S4method{des}{mina}(x)
}
\arguments{
\item{x}{The `mina` object.}

\item{value}{The value to set for the slot of the `mina` object `x`.}
}
\value{
The `des` slot of the `mina` object.
}
\description{
Setter and getter for the slot `des`, which is the description and meta data
of rows in `tab`.
}
\examples{
des(maize) <- maize_des2
head(des(maize))
}
