% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{dis<-}
\alias{dis<-}
\alias{dis}
\alias{dis<-,mina-method}
\alias{dis,mina-method}
\title{Setter and getter for the slot `dis`.}
\usage{
dis(x) <- value

dis(x)

\S4method{dis}{mina}(x) <- value

\S4method{dis}{mina}(x)
}
\arguments{
\item{x}{The `mina` object.}

\item{value}{The value to set for the slot of the `mina` object `x`.}
}
\description{
Setter and getter for the slot `dis`.

Get the slot `dis`
}
