% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{dis_bs}
\alias{dis_bs}
\title{Getter for the slots `dis_bs`, `dis_pm` and `dis_stat`.}
\usage{
dis_bs(x)
}
\arguments{
\item{x}{The `mina` object.}
}
\description{
Getter for the slots `dis_bs`, `dis_pm` and `dis_stat`.
}
