% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_dis.R
\name{dis_par}
\alias{dis_par}
\title{Function for dis_par dissimilarity / distance calculation. Modified from
https://github.com/defleury/Schmidt_et_al_2016_community_similarity/blob/
master/functions.community_similarity.R}
\usage{
dis_par(x, method = "bray", threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An matrix for dissimilarity / distance calculation.}

\item{method}{The method for dissimilarity / distance calculation.}

\item{threads}{(optional) The number of threads used for parallel running,
80 by default.}

\item{nblocks}{(optional) The number of row / column for splitted sub-matrix,
400 by default.}
}
\value{
y The dissimilarity / distance matrix.
}
\description{
Function for dis_par dissimilarity / distance calculation. Modified from
https://github.com/defleury/Schmidt_et_al_2016_community_similarity/blob/
master/functions.community_similarity.R
}
\examples{
y <- dis_par(x, method = "bray", threads = 80, nblocks = 400)
}
\keyword{internal}
