% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{dis_bs}
\alias{dis_bs}
\alias{dis_pm}
\alias{dis_stat}
\title{Getter for the slots `dis_bs`, `dis_pm` and `dis_stat`.}
\usage{
dis_bs(x)

dis_pm(x)

dis_stat(x)
}
\arguments{
\item{x}{The `mina` object.}
}
\value{
The `dis_bs`, `dis_pm` and `dis_stat` slots of the `mina` object.
}
\description{
Getter for the slots `dis_bs`, `dis_pm` and `dis_stat`.
}
\examples{
maize <- new("mina", tab = maize_asv2, des = maize_des2)
maize <- norm_tab(maize, method = "raref")
maize <- fit_tabs(maize)
maize <- bs_pm(maize, group = "Compartment", per = 0.5)
maize <- net_dis(maize, method = "Jaccard")
dis_bs(maize)
dis_pm(maize)
dis_stat(maize)
}
