% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_dis.R
\docType{data}
\name{distanceMethodList}
\alias{distanceMethodList}
\title{List of distance method keys supported in \code{\link{comDis}}}
\format{A list of character vectors.

\describe{
  \item{\code{tina}}{\code{\link[phyloseq]{UniFrac}}}
  \item{\code{manhattan}}{\code{\link[vegan]{vegdist}}}
  \item{\code{euclidean}}{\code{\link[vegan]{vegdist}}}
  \item{\code{canberra}}{\code{\link[vegan]{vegdist}}}
  \item{\code{bray}}{\code{\link[vegan]{vegdist}}}
  \item{\code{kulczynski}}{\code{\link[vegan]{vegdist}}}
  \item{\code{jaccard}}{\code{\link[vegan]{vegdist}}}
  \item{\code{gower}}{\code{\link[vegan]{vegdist}}}
  \item{\code{altGower}}{\code{\link[vegan]{vegdist}}}
  \item{\code{morisita}}{\code{\link[vegan]{vegdist}}}
  \item{\code{horn}}{\code{\link[vegan]{vegdist}}}
  \item{\code{mountford}}{\code{\link[vegan]{vegdist}}}
  \item{\code{raup}}{\code{\link[vegan]{vegdist}}}
  \item{\code{binomial}}{\code{\link[vegan]{vegdist}}}
  \item{\code{chao}}{\code{\link[vegan]{vegdist}}}
  \item{\code{cao}}{\code{\link[vegan]{vegdist}}}
}}
\usage{
distanceMethodList
}
\description{
Add methods can only be matched exactly.
}
\examples{
distanceMethodList
}
\keyword{datasets}
