% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr.R
\docType{methods}
\name{dmr,matrix-method}
\alias{dmr,matrix-method}
\title{Dimensionality reduction of the distance matrix.}
\usage{
\S4method{dmr}{matrix}(x, k = 2)
}
\arguments{
\item{x}{A distance matrix.}

\item{k}{The number of dimensionality after redunction, 2 by default.}
}
\value{
y The coordinates of components indicated in distance matrix in k
dimension.
}
\description{
Dimensionality reduction of the distance matrix.
}
\examples{
y <- dmr(x)
y <- dmr(x, k = 5)
}
