% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_pm.R
\name{filter_mat}
\alias{filter_mat}
\title{Function for filtering of matrix, rows present in less than `p` columns will
be removed. After row filtering, columns with a sum of 0 wil be removed.}
\usage{
filter_mat(x, p)
}
\arguments{
\item{x}{The input matrix to be filtered.}

\item{p}{The cutoff for non-zero column number.}
}
\value{
x The same matrix after filtering.
}
\description{
Function for filtering of matrix, rows present in less than `p` columns will
be removed. After row filtering, columns with a sum of 0 wil be removed.
}
\examples{
data(maize_asv)
maize_asv_filter <- filter_mat(maize_asv, p = 10)
}
\keyword{internal}
