% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{fit_tabs}
\alias{fit_tabs}
\title{Filter the quantitative and descriptive table to make them have the same
samples, the intersect samples will be remained.}
\usage{
fit_tabs(x)
}
\arguments{
\item{x}{An object of the class mina with @tab and @des_tab defined or a
quantitative matrix(need parameter des in this case).}
}
\description{
Filter the quantitative and descriptive table to make them have the same
samples, the intersect samples will be remained.
}
\examples{
fit_tabs(x)
}
