% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_node_cmp.R
\name{get_contrast}
\alias{get_contrast}
\title{Get the contrast between two networks / adjacency matrices.}
\usage{
get_contrast(x, y)
}
\arguments{
\item{x}{The network/adjacency matrix of one of the compared condition.}

\item{y}{The other network/adjacency matrix.}
}
\value{
The contrast between two matrices.
}
\description{
Get the contrast between two networks / adjacency matrices.
}
\keyword{internal}
