% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_grp_cmp.R
\name{get_contrast_grp}
\alias{get_contrast_grp}
\title{Get the contrast of groups of nodes between two networks/adjacency matrices.}
\usage{
get_contrast_grp(x, y, grp)
}
\arguments{
\item{x}{The network/adjacency matrix of one of the compared condition.}

\item{y}{The other network/adjacency matrix.}

\item{grp}{The group information.}
}
\value{
The contrast between two matrices.
}
\description{
Get the contrast of groups of nodes between two networks/adjacency matrices.
}
\keyword{internal}
