% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_dis.R
\name{get_dis_df}
\alias{get_dis_df}
\title{Function for getting distance data frame from `dist`.}
\usage{
get_dis_df(x)
}
\arguments{
\item{x}{The object of class `dist`.}
}
\value{
The data frame of distance matrix.
}
\description{
Function for getting distance data frame from `dist`.
}
\keyword{internal}
