% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_node_cmp.R
\name{get_ja}
\alias{get_ja}
\title{Get the Jaccard distance between two networks / adjacency matrices.}
\usage{
get_ja(x, y)
}
\arguments{
\item{x}{The network/adjacency matrix of one of the compared condition.}

\item{y}{The other network/adjacency matrix.}
}
\value{
The Jaccard distance between two matrices.
}
\description{
Get the Jaccard distance between two networks / adjacency matrices.
}
\keyword{internal}
