% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_node_cmp.R
\name{get_ja0}
\alias{get_ja0}
\title{Get the Jaccard0 distance between two networks / adjacency matrices.}
\usage{
get_ja0(x, y)
}
\arguments{
\item{x}{The network/adjacency matrix of one of the compared condition.}

\item{y}{The other network/adjacency matrix.}
}
\value{
The Jaccard0 distance between two matrices.
}
\description{
Get the Jaccard0 distance between two networks / adjacency matrices.
}
\keyword{internal}
