% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_grp_cmp.R
\name{get_ja0_grp}
\alias{get_ja0_grp}
\title{Get the Jaccard0 distance between two networks / adjacency matrices.}
\usage{
get_ja0_grp(x, y, grp)
}
\arguments{
\item{x}{The network/adjacency matrix of one of the compared condition.}

\item{y}{The other network/adjacency matrix.}

\item{grp}{The table with group information.}
}
\value{
The Jaccard0 distance between two matrices.
}
\description{
Get the Jaccard0 distance between two networks / adjacency matrices.
}
\keyword{internal}
