% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rep.R
\name{get_rep,mina-method}
\alias{get_rep,mina-method}
\title{Get the representative community members.}
\usage{
\S4method{get_rep}{mina}(x, top = 5)
}
\arguments{
\item{x}{An object of the class `mina` with @norm define.}

\item{top}{The percent of the most abundant and prevalent members.}
}
\value{
The same object with @norm replaced by the representative members.
}
\description{
Get the representative community members.
}
\examples{
maize <- new("mina", tab = maize_asv, des = maize_des)
maize <- get_rep(maize)
}
