% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_dis.R
\name{get_stat}
\alias{get_stat}
\title{Function for distance statistic and significance test.}
\usage{
get_stat(x, p = NULL)
}
\arguments{
\item{x}{The bootstrap distance data frame.}

\item{p}{The permuation distance data frame.}
}
\description{
Function for distance statistic and significance test.
}
\examples{
maize <- new("mina", tab = maize_asv2, des = maize_des2)
maize <- norm_tab(maize, method = "raref")
maize <- bs_pm(maize, group = "Compartment")
maize_bs <- maize@multi
maize_pm <- maize@perm
maize_stat <- get_stat(maize_bs, maize_pm)
}
\keyword{internal}
