% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{maize_des2}
\alias{maize_des2}
\title{Subset of design file of maize project, 313 samples are included.}
\format{
A data frame with columns:
\describe{
 \item{Sample_ID}{The unique ID of the microbial profiling sample.}
 \item{Host_genotype}{The genotype of the plant host maize.}
 \item{Compartment}{The compartment of the microbial sample comes from.}
 \item{Soil}{The soil of the sampled microbiome.}
 \item{Management}{The management of the soil where microbial sample from.}
}
}
\source{
RECONSTRUCT project, maize microbiome part.
}
\description{
Subset of design file of maize project, 313 samples are included.
}
\examples{
data(maize_des2)

}
