% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{mat_or_NULL-class}
\alias{mat_or_NULL-class}
\title{Modified from https://github.com/joey711/phyloseq/blob/master/R/allClasses.R
Use setClassUnion to define the unholy NULL-data union as a virtual class.
This is a way of dealing with the expected scenarios in which one or more of
the component data classes is not available, in which case NULL will be used
instead.}
\description{
Modified from https://github.com/joey711/phyloseq/blob/master/R/allClasses.R
Use setClassUnion to define the unholy NULL-data union as a virtual class.
This is a way of dealing with the expected scenarios in which one or more of
the component data classes is not available, in which case NULL will be used
instead.
}
\keyword{internal}
