% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{mina-class}
\alias{mina-class}
\alias{show,}
\alias{mina-method}
\alias{comDis}
\title{Class "mina" includes the quantitative table and descriptive table.}
\description{
Class "mina" includes the quantitative table and descriptive table.
}
\section{Slots}{

\describe{
\item{\code{Tab}}{the quantitative table of the dataset.}

\item{\code{desTab}}{the descriptive table of the samples listed in Tab.}

\item{\code{normTab}}{the normalized quantitative table of Tab.}

\item{\code{dis}}{the distance / dissimilarity matrix between samples in Tab and desTab.}

\item{\code{dmr}}{the list of dimensionality reduction result, includes points and variance.}

\item{\code{adj}}{the adjacency matrix between pairwise compositions (e.g. OTUs/ASVs)}

\item{\code{cls}}{the cluster information for each composition}

\item{\code{clsTab}}{the cluster quantitative table}

\item{\code{disNet}}{the distance between networks of different environmental communities.}

\item{\code{multi}}{the list of subsampled adjacency matrices for each environment.}

\item{\code{perm}}{the list of permutated adjacency matrices for each pairwise environmental comparison.}

\item{\code{sig}}{the average distance between subsampled environmental community networks and corresponding
significance.}
}}

\examples{
new(mina, Tab = "otu_table.txt", desTab = "design.txt")
}
\author{
Rui Guan \url{https://github.com/Guan06}
}
