% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_cls.R
\docType{methods}
\name{net_cls,matrix,ANY-method}
\alias{net_cls,matrix,ANY-method}
\title{Network clustering based on the sparsed adjacacency matrix @adj.}
\usage{
\S4method{net_cls}{matrix,ANY}(x, method, cutoff = 0.4, neg = FALSE)
}
\arguments{
\item{x}{Adjacency matrix used for clustering.}

\item{method}{The clustering method used.}

\item{cutoff}{The cutoff for the sparsed adjacacency matrix, default 0.4.}

\item{neg}{Whether to keep the negative edges, cannot be TRUE when using
`mcl` for clustering. Default FALSE.}
}
\value{
y The cluster table.
}
\description{
Network clustering based on the sparsed adjacacency matrix @adj.
}
\examples{
y <- net_cls(x, method = "mcl")
y <- net_cls(x, method = "mcl", cutoff = 0.4, neg = FALSE)
}
