% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_cls.R
\docType{methods}
\name{net_cls,mina,ANY-method}
\alias{net_cls,mina,ANY-method}
\title{Network clustering based on the sparsed adjacacency matrix @adj.}
\usage{
\S4method{net_cls}{mina,ANY}(x, method, cutoff = 0.4, neg = FALSE)
}
\arguments{
\item{x}{An object of class `mina` with @adj defined.}

\item{method}{The clustering method used.}

\item{cutoff}{The cutoff for the sparsed adjacacency matrix, default 0.4.}

\item{neg}{Whether to keep the negative edges, cannot be TRUE when using
`mcl` for clustering. Default FALSE.}
}
\value{
x The same `mina` class with @cls added.
}
\description{
Network clustering based on the sparsed adjacacency matrix @adj.
}
\examples{
data(maize)
maize <- norm_tab(maize, method = "raref")
maize <- adj(maize, method = "spearman")
maize@adj <- maize@adj[1:500, 1:500]
maize <- net_cls(maize, method = "mcl")
}
