% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_dis.R
\docType{methods}
\name{net_dis,mina,ANY-method}
\alias{net_dis,mina,ANY-method}
\title{Calculate the network distance of @multi and test the significance when @perm
is defined.}
\usage{
\S4method{net_dis}{mina,ANY}(x, method, sig = TRUE)
}
\arguments{
\item{x}{An object of class `mina` with @multi (and @perm if sig is TRUE)
defined.}

\item{method}{The distance to be calculated, "spectral" and "jaccard" are
available.}

\item{sig}{Whether to test the significance, if TRUE (by default), @perm is
needed.}
}
\value{
x The same `mina` object with @net_dis defined.
}
\description{
Calculate the network distance of @multi and test the significance when @perm
is defined.
}
\examples{
x <- net_dis(x, method = "spectral", sig = TRUE)
}
