% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_dis.R
\docType{methods}
\name{net_dis,mina,ANY-method}
\alias{net_dis,mina,ANY-method}
\alias{net_dis,mina,character-method}
\title{Calculate the network distance of @multi and test the significance when @perm
is defined.}
\usage{
\S4method{net_dis}{mina,ANY}(x, method, sig = TRUE)

\S4method{net_dis}{mina,character}(x, method, sig = TRUE)
}
\arguments{
\item{x}{An object of class `mina` with @multi (and @perm if sig is TRUE)
defined.}

\item{method}{The distance to be calculated, "spectral" and "jaccard" are
available.}

\item{sig}{Whether to test the significance, if TRUE (by default), @perm is
needed.}
}
\value{
x The same `mina` object with @net_dis defined.
}
\description{
Calculate the network distance of @multi and test the significance when @perm
is defined.
}
\examples{
\dontrun{
data(maize)
maize <- norm_tab(maize, method = "raref")
maize <- bs_pm(maize, group = "Compartment")
maize <- net_dis(maize)
}
}
