% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{net_dis}
\alias{net_dis}
\title{Calculate the network distance of @multi and test the significance when @perm
is defined.}
\usage{
net_dis(x, method, sig = TRUE)
}
\arguments{
\item{x}{An object of class `mina` with @multi (and @perm if sig is TRUE)
defined.}

\item{method}{The distance to be calculated, "spectral" and "jaccard" are
available.}

\item{sig}{Whether to test the significance, if TRUE (by default), @perm is
needed.}
}
\description{
Calculate the network distance of @multi and test the significance when @perm
is defined.
}
\examples{
x <- net_dis(x, method = "spectral", sig = TRUE)
}
