% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{dis_bs<-}
\alias{dis_bs<-}
\alias{dis_pm<-}
\alias{dis_stat<-}
\alias{dis_bs<-,mina-method}
\alias{dis_pm<-,mina-method}
\alias{dis_stat<-,mina-method}
\title{Setter for the slots 'dis_bs', `dis_pm` and `dis_stat`.}
\usage{
dis_bs(x) <- value

dis_pm(x) <- value

dis_stat(x) <- value

\S4method{dis_bs}{mina}(x) <- value

\S4method{dis_pm}{mina}(x) <- value

\S4method{dis_stat}{mina}(x) <- value
}
\arguments{
\item{x}{The `mina` object.}

\item{value}{The value to set for the slot of the `mina` object `x`.}
}
\value{
The `dis_bs`, `dis_pm` and `dis_stat` slots of the `mina` object.
}
\description{
Setter for the slots 'dis_bs', `dis_pm` and `dis_stat`.
}
\keyword{internal}
