% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/net_dis_pcoa.R
\name{net_dis_pcoa}
\alias{net_dis_pcoa}
\alias{net_dis_pcoa,character-method}
\title{Visulization of spectra network distance as PCoA.}
\usage{
net_dis_pcoa(x)

\S4method{net_dis_pcoa}{character}(x)
}
\arguments{
\item{x}{The folder with all egv files generated by net_dis_indi().}
}
\value{
p The plotted figure.

p The plotted figure.
}
\description{
Visulization of spectra network distance as PCoA.

Visulization of spectra network distance as PCoA.
}
\examples{
\dontrun{
data(maize)
norm(maize) <- maize_asv2
maize <- fit_tabs(maize)
maize <- bs_pm(maize, group = "Compartment", individual = TRUE, out_dir =
"./individual_bs_pm/")
maize <- net_dis_indi("./individual_bs_pm/", method = "spectra", egv = TRUE,
dir = "./egv_folder/")
p <- net_dis_pcoa("./egv_folder/")
}
\dontrun{
data(maize)
maize <- norm_tab(maize)
maize <- fit_tabs(maize)
maize <- get_rep(maize, top = 5)
maize <- bs_pm(maize, group = "Compartment", individual = TRUE, out_dir =
"./individual_bs_pm/")
maize <- net_dis_indi("./individual_bs_pm/", method = "spectra", egv = TRUE,
dir = "./egv_folder/")
p <- net_dis_pcoa("./egv_folder/")
}
}
