% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/net_dis_plot.R
\name{net_dis_plot}
\alias{net_dis_plot}
\alias{net_dis_plot,mina-method}
\title{Visulization of network distance, average distances are used for tile plot.}
\usage{
net_dis_plot(x, d = "BS", ...)

\S4method{net_dis_plot}{mina}(x, d = "BS", sig = TRUE)
}
\arguments{
\item{x}{An object of `mina` with slot `dis_stat` defined.}

\item{d}{The distance to be plotted, could be "BS" or "PM".}

\item{...}{Additional parameters.}

\item{sig}{If `TRUE`, indicating significant distance with gold guild.}
}
\value{
p The plotted figure.

p The plotted figure.
}
\description{
Visulization of network distance, average distances are used for tile plot.

Visulization of network distance, average distances are used for tile plot.
}
\examples{
maize <- new("mina", tab = maize_asv2, des = maize_des2)
maize <- norm_tab(maize, method = "raref")
maize <- fit_tabs(maize)
maize <- bs_pm(maize, group = "Compartment")
maize <- net_dis(maize, method = "Jaccard")
p <- net_dis_plot(maize)
maize <- new("mina", tab = maize_asv2, des = maize_des2)
maize <- norm_tab(maize, method = "raref")
maize <- fit_tabs(maize)
maize <- bs_pm(maize, group = "Compartment")
maize <- net_dis(maize, method = "Jaccard")
p <- net_dis_plot(maize, d = "BS")
}
