% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/net_grp_cmp.R
\name{net_grp_cmp}
\alias{net_grp_cmp}
\alias{net_grp_cmp,character-method}
\title{Compare the group features between networks.}
\usage{
net_grp_cmp(x, cmp = "contrast", dir = "./", grp)

\S4method{net_grp_cmp}{character}(x, cmp = "contrast", dir = "./", grp)
}
\arguments{
\item{x}{The folder with all network inference results generated by bs_pm()}

\item{cmp}{The compared feature of grp, default `contrast`.}

\item{dir}{The directory to store the alculated node features.}

\item{grp}{The table with group information.}
}
\description{
Compare the group features between networks.

Compare the group features between networks.
}
\examples{
\dontrun{
net_node_cmp("./individual_bs_pm/", f = "contrast", dir = "./", grp =
cls_tab(maize))
}
\dontrun{
net_node_cmp("./individual_bs_pm/", f = "contrast", dir = "./", grp =
cls_tab(maize))
}
}
