% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/net_node_cmp.R
\name{net_node_cmp}
\alias{net_node_cmp}
\alias{net_node_cmp,character,character-method}
\title{Compare the node features between networks.}
\usage{
net_node_cmp(x, cmp = "contrast", dir = "./")

\S4method{net_node_cmp}{character,character}(x, cmp = "contrast", dir = "./")
}
\arguments{
\item{x}{The folder with all network inference results generated by bs_pm()}

\item{cmp}{The compared feature of node, default `contrast`.}

\item{dir}{The directory to store the alculated node features.}
}
\description{
Compare the node features between networks.

Compare the node features between networks.
}
\examples{
\dontrun{
net_node_cmp("./individual_bs_pm/", f = "contrast", dir = "./")
}
\dontrun{
net_node_cmp("./individual_bs_pm/", f = "contrast", dir = "./")
}
}
