% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{norm<-}
\alias{norm<-}
\alias{norm<-,mina-method}
\alias{norm}
\alias{norm,mina-method}
\title{Setter and getters for the slot `norm`, normalized `tab` matrix.}
\usage{
norm(x) <- value

\S4method{norm}{mina}(x) <- value

norm(x)

\S4method{norm}{mina}(x)
}
\arguments{
\item{x}{The `mina` object.}

\item{value}{The value to set for the slot of the `mina` object `x`.}
}
\description{
Setter and getters for the slot `norm`, normalized `tab` matrix.
}
\examples{
norm(maize) <- norm_tab(maize_asv2, method = "total")
norm(maize)[1:5, 1:5]
}
