% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_tab.R
\name{norm_by_raref}
\alias{norm_by_raref}
\title{Function for normalization by rarefying the samples into the same depth,
modified from \pkg{phyloseq}, find it
\href{https://rdrr.io/bioc/phyloseq/man/rarefy_even_depth.html}{here}.}
\usage{
norm_by_raref(x, depth = 1000, replace = TRUE)
}
\arguments{
\item{x}{A quantitative table with sample in columns and compositions in rows.}

\item{depth}{The depth for rarefying, 1000 by default.}

\item{replace}{Whether to sample with replacement (\code{TRUE}) or without
replacement (\code{FALSE}). Default \code{TRUE} for computational efficiency.}
}
\value{
A normalized quantitative table.
}
\description{
Function for normalization by rarefying the samples into the same depth,
modified from \pkg{phyloseq}, find it
\href{https://rdrr.io/bioc/phyloseq/man/rarefy_even_depth.html}{here}.
}
\keyword{internal}
