% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_tab.R
\name{norm_by_total}
\alias{norm_by_total}
\title{Function for normalization, by total number of the reads in each sample.}
\usage{
norm_by_total(x)
}
\arguments{
\item{x}{A quantitative table with samples in cloumns and compostions in
rows.}
}
\value{
A normalized quantitative table.
}
\description{
Function for normalization, by total number of the reads in each sample.
}
\examples{
norm_tab <- norm_by_total(mina@tab)
}
\keyword{internal}
