% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_tab.R
\docType{methods}
\name{norm_tab,matrix,character-method}
\alias{norm_tab,matrix,character-method}
\title{Normalize the quantitative matrix.}
\usage{
\S4method{norm_tab}{matrix,character}(x, method, depth = 1000,
  replace = TRUE)
}
\arguments{
\item{x}{A quantitative matrix with samples in columns and compostions in
rows.}

\item{method}{The method used for normalization.}

\item{depth}{The depth for rarefying, 1000 by default.}

\item{replace}{Whether to sample with replacement (\code{TRUE} by default)
or without replacement (\code{FALSE}) when using method `raref`.}
}
\value{
x_norm Normalized matrix of the quantitative table.
}
\description{
Normalize the quantitative matrix.
}
\examples{
data(maize_asv)
maize_asv_norm <- norm_tab(maize_asv, method = "total")
maize_asv_norm <- norm_tab(maize_asv, method = "raref", depth = 1000, replace
= TRUE)
}
