% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{norm_tab}
\alias{norm_tab}
\title{Normalize the slot `tab` for later analysis.}
\usage{
norm_tab(x, method, ...)
}
\arguments{
\item{x}{The input mina object with quantitative tab / a matrix needed to be
normalized.}

\item{method}{The method used for the normalization of quantitative table.}

\item{...}{Additional parameters.}
}
\value{
Normalized quantitative table.
}
\description{
Normalize the slot `tab` for later analysis.
}
\examples{
maize <- new("mina", tab = maize_asv2, des = maize_des2)
maize <- norm_tab(maize, method = "total")
}
