% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_tab.R
\name{rarefaction_subsample}
\alias{rarefaction_subsample}
\title{Rarefaction subsample function, one sample, modified from a internal function
in \pkg{phyloseq}, find it
\href{https://rdrr.io/bioc/phyloseq/man/rarefy_even_depth.html}{here}.}
\usage{
rarefaction_subsample(x, depth = 1000, replace = TRUE)
}
\arguments{
\item{x}{A column of quantitative table.}

\item{depth}{The depth for rarefying, 1000 by default.}

\item{replace}{Whether to sample with or without replacement, \code{TRUE} by
default for computational efficiency.}
}
\value{
The vector rarefied to defined depth.
}
\description{
Rarefaction subsample function, one sample, modified from a internal function
in \pkg{phyloseq}, find it
\href{https://rdrr.io/bioc/phyloseq/man/rarefy_even_depth.html}{here}.
}
\keyword{internal}
