% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{tab<-}
\alias{tab<-}
\alias{tab<-,mina-method}
\alias{tab}
\alias{tab,mina-method}
\title{Setter and getter for the slot `tab`.}
\usage{
tab(x) <- value

\S4method{tab}{mina}(x) <- value

tab(x)

\S4method{tab}{mina}(x)
}
\arguments{
\item{x}{The `mina` object.}

\item{value}{The value to set for the slot of the `mina` object `x`.}
}
\description{
Setter and getter for the slot `tab`.
}
\examples{
tab(maize) <- maize_asv2
tab(maize)[1:5, 1:5]
}
