% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{tina}
\alias{tina}
\title{TINA community dissimilarity used in \code{\link[mina]{com_dis}}.
Function for `tina` dissimilarity/distance calculation. Modified from Schmidt
et al., 2016.}
\usage{
tina(
  x,
  cor_method = "spearman",
  sim_method = "w_ja",
  threads = 80,
  nblocks = 400
)
}
\arguments{
\item{x}{An matrix for dissimilarity calculation.}

\item{cor_method}{The method for correlation, "pearson" and "spearman" are
available.}

\item{sim_method}{The method for similarity, "w_ja" and "uw_ja" are
available for weighted and unweighted Jaccard similarity respectively.}

\item{threads}{The number of threads used for parallel running, 80 by
default.}

\item{nblocks}{The number of row / column for splitted sub-matrix, 400 by
default.}
}
\description{
TINA community dissimilarity used in \code{\link[mina]{com_dis}}.
Function for `tina` dissimilarity/distance calculation. Modified from Schmidt
et al., 2016.
}
\examples{
\dontrun{
data(maize)
maize@tab <- maize@tab[1 : 1000, 1 : 200]
maize <- norm_tab(maize, method = "raref", depth = 100)
maize <- fit_tabs(maize)
asv_norm <- maize@norm
asv_tina <- tina(asv_norm, cor_method = "spearman", sim_method = "w_ja",
threads = 2, nblocks = 40)
}
}
