% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_des}
\alias{add_des}
\title{Add tables to the object}
\usage{
add_des(mina, file)
}
\arguments{
\item{mina}{An object of the class mina with Tab defined.}

\item{file}{The descriptive table.}
}
\value{
mina An object of the class mina with descriptive table added.
}
\description{
This includes functions for adding descriptive table (e.g. design file)
to mina object for later analysis. The samples in first column "SampleID" 
should be the same as the samples in the quantitative table.
}
\examples{
add_des(mina, "design.txt")

}
