% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_tab}
\alias{add_tab}
\title{Add quantitative table to the object.}
\usage{
add_tab(x, file)
}
\arguments{
\item{x}{An object of the class mina.}

\item{file}{The table to be added to the object.}
}
\value{
x An object of the class mina with @tab added.
}
\description{
This includes functions for adding quantitative table (e.g. OTU table) to
mina object for later analysis.
}
\examples{
add_tab(x, "otu_table.txt")
add_tab(x, "asv_table.rds")

}
