% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addDes}
\alias{addDes}
\title{Add tables to the object}
\usage{
addDes(mina, file)
}
\arguments{
\item{mina}{An object of the class mina.}

\item{file}{The descriptive table.}
}
\value{
mina with descriptive table.
}
\description{
This includes functions for adding descriptive table (e.g. design file)
to mina object for later analysis.
}
\examples{
addTab(mina, "design.txt")

}
