% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\docType{methods}
\name{adj-matrix}
\alias{adj-matrix}
\alias{adj,matrix,character-method}
\title{Calculate the adjacacency matrix of @norm by correlation with matrix as
input.}
\usage{
\S4method{adj}{matrix,ANY}(x, method, threads = 80, nblocks = 400)

\S4method{adj}{matrix,character}(x, method, threads = 80,
  nblocks = 400)
}
\arguments{
\item{x}{An matrix for correlation / adjacency matrix calculation.}

\item{method}{The correlation coefficient used for adjacacency matrix.}

\item{threads}{The number of threads used for parallel running, 80 by
default.}

\item{nblocks}{The number of row / column for splitted sub-matrix, 400 by
default.}
}
\value{
y The adjacacency matrix.
}
\description{
Calculate the adjacacency matrix of @norm by correlation with matrix as
input.
}
\examples{
\dontrun{
data(maize_asv)
maize_asv <- maize_asv[1:500, 1:300]
maize_asv_adj <- adj(maize_asv, method = "sparcc", threads = 8, nblocks = 40)
}
}
