% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\name{adj,mina,ANY-method}
\alias{adj,mina,ANY-method}
\alias{adj,mina,character-method}
\title{Calculate the adjacency matrix of @norm by correlation with `mina` class
object as input.}
\usage{
\S4method{adj}{mina,ANY}(x, method, sig = FALSE, threads = 80, nblocks = 400)

\S4method{adj}{mina,character}(x, method, sig = FALSE, threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An object of the class `mina` with @norm defined.}

\item{method}{The correlation coefficient used for adjacency matrix.}

\item{sig}{The asymtotic P-values, only applicable for Pearson and Spearman
methods, FALSE by default.}

\item{threads}{The number of threads used for parallel running, 80 by
default.}

\item{nblocks}{The number of row/column for splitting sub-matrix, 400 by
default.}
}
\value{
x The same `mina` object with @adj added.
}
\description{
Calculate the adjacency matrix of @norm by correlation with `mina` class
object as input.
}
\examples{
\dontrun{
data(maize)
maize@tab <- maize@tab[1 : 500, 1 : 300]
maize <- norm_tab(maize, method = "raref")
maize <- fit_tabs(maize)
maize <- adj(maize, method = "pearson")
maize <- adj(maize, method = "spearman")
maize <- adj(maize, method = "spearman", sig = FALSE)
maize <- adj(maize, method = "sparcc", threads = 2, nblocks = 40)
}
}
