% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{adj}
\alias{adj}
\title{Calculate the correlation adjacacency matrix.}
\usage{
adj(x, method, ...)
}
\arguments{
\item{x}{An object of the class mina with `norm` defined or a `norm` matrix.}

\item{method}{The correlation coeffient used for adjacacency matrix.}

\item{...}{Additional parameters.}
}
\value{
Adjacency matrix.
}
\description{
Calculate the correlation adjacacency matrix.
}
\examples{
maize <- new("mina", tab = maize_asv2, des = maize_des2)
maize <- norm_tab(maize, method = "raref", depth = 1000)
maize <- fit_tabs(maize)
maize <- adj(maize, method = "spearman", sig = FALSE)
}
