% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{adj}
\alias{adj}
\title{Calculate the adjacacency matrix of @norm by correlation.}
\usage{
adj(x, method, sig = FALSE, threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An object of the class mina with @norm defined.}

\item{method}{The correlation coeffient used for adjacacency matrix.}

\item{sig}{(optional) The asymtotic P-values, only applicable for Pearson and
Spearman methods.}

\item{threads}{(optional) The number of threads used for parallel running.}

\item{nblocks}{(optional) The number of row / column for splitted sub-matrix.}
}
\value{
Adjacency matrix.
}
\description{
Calculate the adjacacency matrix of @norm by correlation.
}
\examples{
maize <- new("mina", tab = maize_asv2, des = maize_des2)
maize <- norm_tab(maize, method = "raref", depth = 1000)
maize <- fit_tabs(maize)
maize <- adj(maize, method = "spearman")
maize <- adj(maize, method = "spearman", sig = FALSE)
}
