% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{adj}
\alias{adj}
\title{Calculate the adjacacency matrix of @norm by correlation.}
\usage{
adj(x, method, threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An object of the class mina with @norm defined.}

\item{method}{The correlation coeffient used for adjacacency matrix.}

\item{threads}{(optional) The number of threads used for parallel running.}

\item{nblocks}{(optional) The number of row / column for splitted sub-matrix.}
}
\description{
Calculate the adjacacency matrix of @norm by correlation.
}
\examples{
data(maize)
maize@tab <- maize@tab[1 : 500, 1 : 200]
maize <- norm_tab(maize, method = "raref", depth = 200)
maize <- fit_tabs(maize)
maize <- adj(maize, method = "spearman")
maize <- adj(maize, method = "sparcc", threads = 2, nblocks = 40)
}
