% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{.adj<-}
\alias{.adj<-}
\alias{.adj<-,mina-method}
\alias{.adj}
\alias{adj_sig<-}
\alias{adj_sig<-,mina-method}
\title{Setter for the slot `adj` and `adj_sig`, the adjacency matrix of
`norm` and corresponding significant value matrix with `sig` is `TRUE`.}
\usage{
.adj(x) <- value

\S4method{.adj}{mina}(x) <- value

.adj(x)

adj_sig(x) <- value

\S4method{adj_sig}{mina}(x) <- value
}
\arguments{
\item{x}{The `mina` object.}

\item{value}{The value to set for the slot of the `mina` object `x`.}
}
\description{
Setter for the slot `adj` and `adj_sig`, the adjacency matrix of
`norm` and corresponding significant value matrix with `sig` is `TRUE`.

Get the slot `adj`.
}
\examples{
norm(maize) <- norm_tab(maize_asv2, method = "total")
maize <- adj(maize, method = "spearman", sig = TRUE)
adj_sig(maize)[1:5, 1:5]
}
\keyword{internal}
