% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{bs_pm}
\alias{bs_pm}
\title{Inferring the network of different group of samples and test significance by
permutation.}
\usage{
bs_pm(x, group, g_size = 100, s_size = 50, rm = TRUE, sig = TRUE,
  bs = 6, pm = 6)
}
\arguments{
\item{x}{An object of class `mina` with @norm and @des defined.}

\item{group}{The column name of descriptive file @des for comparison.}

\item{g_size}{The cutoff of group size used for filtering, default 100.}

\item{s_size}{The number of samples used for network inference during
bootstrap and permutation (when sig == TRUE), it should be smaller than
g_size to make sure the randomness; default 50.}

\item{rm}{Filtering the components present in less than 20% of the samples,
default TRUE.}

\item{sig}{Whether to test the significance, skip the permutation when sig ==
FALSE, default TRUE.}

\item{bs}{The times for bootstrap network inference, default 6}

\item{pm}{The times for permuatated samples network inference, default 6.}
}
\description{
Inferring the network of different group of samples and test significance by
permutation.
}
\examples{
x <- bs_pm(x, group = "Compartment")
x <- bs_pm(x, group = "Compartment", g_size = 100, s_size = 50, rm = TRUE,
sig = TRUE, bs = 6, pm = 6)
}
