% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mina}
\alias{check_mina}
\title{Check the tab and des_tab file. Return TRUE if they are NULL or both
quantitative and descriptive files of same samples are included (i.e. the
object is valid).}
\usage{
check_mina(x)
}
\arguments{
\item{x}{An object of class mina.}
}
\value{
TRUE if the object is valid or empty.
}
\description{
Check the tab and des_tab file. Return TRUE if they are NULL or both
quantitative and descriptive files of same samples are included (i.e. the
object is valid).
}
\examples{
check_mina(x)
}
