% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mina_qu}
\alias{check_mina_qu}
\title{Check the object and return TRUE if the object includes quantitative table.}
\usage{
check_mina_qu(object)
}
\arguments{
\item{mina}{a mina object.}
}
\value{
TRUE if the object contains quantitative table.
}
\description{
Check the object and return TRUE if the object includes quantitative table.
}
\examples{
check_mina_qu(mina)
export
}
