% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_dis.R
\docType{methods}
\name{com_dis,matrix,ANY-method}
\alias{com_dis,matrix,ANY-method}
\alias{com_dis,matrix,character-method}
\title{Calculate the community dissimilarity / distance matrix of the input matrix.}
\usage{
\S4method{com_dis}{matrix,ANY}(x, method, threads = 80, nblocks = 400)

\S4method{com_dis}{matrix,character}(x, method, threads = 80,
  nblocks = 400)
}
\arguments{
\item{x}{A matrix of the quantitative table.}

\item{method}{The dissimilarity / distance method used.}

\item{threads}{(optional, only needed when method == "tina") The number of
threads used for parallel running.}

\item{nblocks}{(optional, only needed when method == "tina") The number of
row / column for splitted sub-matrix.}
}
\value{
y The dissimilarity / distance matrix.
}
\description{
Calculate the community dissimilarity / distance matrix of the input matrix.
}
\examples{
\dontrun{
data(maize)
maize@tab <- maize@tab[1 : 1000, 1 : 200]
maize <- norm_tab(maize, method = "raref", depth = 100)
maize <- fit_tabs(maize)
asv_norm <- maize@norm
asv_dis <- com_dis(asv_norm, method = "bray")
}
}
