% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{com_dis}
\alias{com_dis}
\title{Calculate the community dissimilarity / distance matrix of @norm.}
\usage{
com_dis(x, method, threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An object of the class mina with @norm defined or any quantitative
matrix.}

\item{method}{The dissimilarity / distance method used.}

\item{threads}{(optional) The number of threads used for parallel running.}

\item{nblocks}{(optional) The number of row / column for splitted sub-matrix.}
}
\description{
Calculate the community dissimilarity / distance matrix of @norm.
}
\examples{
\dontrun{
data(maize)
maize <- norm_tab(maize, method = "total")
maize <- com_dis(maize, method = "bray")
}
}
