% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_dis.R
\docType{data}
\name{com_dis_list}
\alias{com_dis_list}
\title{List of dissimilarity / distance supported in \code{\link[mina]{com_dis}}}
\format{A list of character vectors indicate the dissimilarity / distance
method used.

\describe{
  \item{\code{tina}}{TINA from Schmidt_et_al_2016}

  \item{\code{manhattan}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{euclidean}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{canberra}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{bray}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{kulczynski}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{jaccard}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{gower}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{altGower}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{morisita}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{horn}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{mountford}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{raup}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{binomial}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{chao}}{ from \code{\link[vegan]{vegdist}}}
  \item{\code{cao}}{ from \code{\link[vegan]{vegdist}}}
}}
\usage{
com_dis_list
}
\description{
Dissimilarity / distance should be specified by exact string match.
}
\examples{
com_dis_list
}
\keyword{datasets}
