% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_plot.R
\docType{methods}
\name{com_plot,mina,ANY,ANY-method}
\alias{com_plot,mina,ANY,ANY-method}
\title{Visulization of components distance / dissimilarity in k dimension.}
\usage{
\S4method{com_plot}{mina,ANY,ANY}(x, match, color, shape = NULL)
}
\arguments{
\item{x}{An object of `mina` with list @dmr defined.}

\item{match}{The column name of the components IDs in @des with exactly the
same as rownames in x.}

\item{color}{The column name in @des to be used for different color groups.}

\item{shape}{The column name in @des to be used for different shape groups,
default `NULL`.
shape groups.}
}
\value{
p The plotted figure.
}
\description{
Visulization of components distance / dissimilarity in k dimension.
}
\examples{
data(maize)
maize <- norm_tab(maize, method = "raref")
maize <- com_dis(maize, method = "bray")
maize@des <- maize@des[maize@des$Sample_ID \%in\% rownames(maize@dis), ]
maize <- dmr(maize)
p <- com_plot(maize, match = "Sample_ID", color = "Compartment",
              shape = "Soil")
}
