% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cp_cor}
\alias{cp_cor}
\title{Function for correlation coefficient calculation.}
\usage{
cp_cor(mat)
}
\arguments{
\item{mat}{The input matrix for correlation calculation.}
}
\value{
The correlation matrix.
}
\description{
Function for correlation coefficient calculation.
}
