% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr.R
\docType{methods}
\name{dmr,mina-method}
\alias{dmr,mina-method}
\title{Dimensionality reduction of the @dis included in mina.}
\usage{
\S4method{dmr}{mina}(x, k = 2)
}
\arguments{
\item{x}{An object of the class `mina` with @dis defined.}

\item{k}{The number of dimensionality after redunction, 2 by default.}
}
\value{
x The same object with @dmr added.
}
\description{
Dimensionality reduction of the @dis included in mina.
}
\examples{
data(maize)
maize <- norm_tab(maize, method = "raref")
maize@nomm <- maize@norm[1:500, 1:300]
maize <- com_dis(maize, method = "bray", threads = 8, nblocks = 40)
maize <- dmr(maize)
}
