% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{dmr}
\alias{dmr}
\title{Dimensionality reduction of community dissimilarity / distance for
visulization.}
\usage{
dmr(x, k = 2)
}
\arguments{
\item{x}{An object of class `mina` with @dis defined.}

\item{k}{The dimension number after reduction.}
}
\description{
Dimensionality reduction of community dissimilarity / distance for
visulization.
}
\examples{
x <- dmr(x, k = 2)
}
