% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{dmr}
\alias{dmr}
\title{Dimensionality reduction of community dissimilarity / distance for
visulization.}
\usage{
dmr(x, k = 2)
}
\arguments{
\item{x}{An object of class `mina` with @dis defined or a distance matrix.}

\item{k}{The dimension number after reduction.}
}
\value{
The dimentionality reduction results.
}
\description{
Dimensionality reduction of community dissimilarity / distance for
visulization.
}
\examples{
maize <- new("mina", tab = maize_asv2, des = maize_des2)
maize <- norm_tab(maize, method = "raref")
maize <- fit_tabs(maize)
maize <- com_dis(maize, method = "bray")
maize <- dmr(maize)
}
