% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_accessors.R
\name{.dmr<-}
\alias{.dmr<-}
\alias{.dmr}
\alias{.dmr<-,mina-method}
\alias{.dmr,mina-method}
\title{Setter and getter for the slot `dmr`.}
\usage{
.dmr(x) <- value

.dmr(x)

\S4method{.dmr}{mina}(x) <- value

\S4method{.dmr}{mina}(x)
}
\arguments{
\item{x}{The `mina` object.}

\item{value}{The value to set for the slot of the `mina` object `x`.}
}
\description{
Setter and getter for the slot `dmr`.
}
\keyword{internal}
