% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tabs.R
\docType{methods}
\name{fit_tabs,ANY-method}
\alias{fit_tabs,ANY-method}
\title{Filter the quantitative and descriptive table to make them have the same
samples, samples present in both tables are remained. If @norm table exist in
the `mina` object, descriptive table will be filtered again to only keep
samples present in @norm.}
\usage{
\S4method{fit_tabs}{ANY}(x)
}
\arguments{
\item{x}{An object of class mina.}
}
\value{
x The same object as input with fitted @tab and @des_tab.
}
\description{
Filter the quantitative and descriptive table to make them have the same
samples, samples present in both tables are remained. If @norm table exist in
the `mina` object, descriptive table will be filtered again to only keep
samples present in @norm.
}
\examples{
x <- fit_tabs(x)
}
