% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{get_r2}
\alias{get_r2}
\title{Function for unexplained variance ratio calculation indicated in
Anderson, M.J. 2001. A new method for non-parametric multivariate analysis of
variance. Austral Ecology, 26: 32--46.}
\usage{
get_r2(x, des, group = c("Host_genotype", "Compartment", "Soil",
  "Management"))
}
\arguments{
\item{x}{Dissimilarity / distance matrix which indicate variances.}

\item{des}{The descriptive table of samples which define the groups.}

\item{group}{The name(s) of column(s) used  as experimental setup group(s) in
descriptive file.}
}
\value{
r2 The variance ratio cannot be explained by given groups.
}
\description{
Function for unexplained variance ratio calculation indicated in
Anderson, M.J. 2001. A new method for non-parametric multivariate analysis of
variance. Austral Ecology, 26: 32--46.
}
\examples{
r2 <- get_r2(x, des, group = c("Compartment", "Soil")

}
