% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_dis.R
\name{get_spectra}
\alias{get_spectra}
\title{Function for calculation of eigenvalue of given matrix.}
\usage{
get_spectra(x, k = 100)
}
\arguments{
\item{x}{The input matrix.}

\item{k}{Get the first k eigenvalues.}
}
\value{
y The vector of the first k eigenvalues.
}
\description{
Function for calculation of eigenvalue of given matrix.
}
\examples{
\dontrun{
data(maize)
maize@tab <- maize@tab[1 : 500, 1 : 300]
maize <- norm_tab(maize, method = "raref")
maize <- fit_tabs(maize)
maize <- adj(maize, method = "spearman")
maize_spectra <- get_spectra(maize@adj)
}
}
\keyword{internal}
