% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_dis.R
\name{get_spectra}
\alias{get_spectra}
\title{Function for calculation of eigenvalue of given matrix.}
\usage{
get_spectra(x, k = 100)
}
\arguments{
\item{k}{Get the first k eigenvalues.}
}
\value{
y The vector of the first k eigenvalues.
}
\description{
Function for calculation of eigenvalue of given matrix.
}
\examples{
\dontrun{
data(maize)
maize <- norm_tab(maize, method = "raref")
maize <- adj(maize)
y <- get_spectra(maize@adj)
}
}
\keyword{internal}
