% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_dis.R
\name{get_stat}
\alias{get_stat}
\title{Function for distance statistic and significance test.}
\usage{
get_stat(x, p = NULL)
}
\arguments{
\item{x}{The bootstrap distance data frame.}

\item{p}{The permuation distance data frame.}

\item{sig}{Whether to test significance or not.}
}
\description{
Function for distance statistic and significance test.
}
\examples{
y <- get_stat(x)
y <- get_stat(x, p)
}
\keyword{internal}
