% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{mina-class}
\alias{mina-class}
\title{Class "mina" includes the quantitative table and descriptive table.}
\description{
Class "mina" includes the quantitative table and descriptive table.
}
\examples{
new(mina)
}
