% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{net_cls}
\alias{net_cls}
\title{Network clustering of sparsed adjacacency matrix @adj.}
\usage{
net_cls(x, method, cutoff = 0.4, neg = FALSE)
}
\arguments{
\item{x}{An object of class `mina` with @adj defined.}

\item{method}{The clustering method used.}

\item{cutoff}{The cutoff for the sparsed adjacacency matrix, default 0.4.}

\item{neg}{Whether to keep the negative edges, default FALSE.}
}
\description{
Network clustering of sparsed adjacacency matrix @adj.
}
\examples{
\dontrun{
data(maize)
maize@tab <- maize@tab[1 : 1000, 1 : 500]
maize <- norm_tab(maize, method = "raref", depth = 100)
maize <- fit_tabs(maize)
maize <- adj(maize, method = "spearman")
maize <- net_cls(maize, method = "mcl")
maize <- net_cls(maize, method = "mcl", cutoff = 0.4, neg = FALSE)
maize <- net_cls(maize, method = "ap")
maize <- net_cls(maize, method = "ap", cutoff = 0.4, neg = FALSE)
}
}
