% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{net_cls}
\alias{net_cls}
\title{Network clustering of sparsed adjacacency matrix @adj.}
\usage{
net_cls(x, method, cutoff = 0.4, neg = FALSE)
}
\arguments{
\item{x}{An object of class `mina` with @adj defined.}

\item{method}{The clustering method used.}

\item{cutoff}{The cutoff for the sparsed adjacacency matrix, default 0.4.}

\item{neg}{Whether to keep the negative edges, default FALSE.}
}
\description{
Network clustering of sparsed adjacacency matrix @adj.
}
\examples{
x <- net_cls(x)
x <- net_cls(x, method = "mcl", cutoff = 0.4, neg = FALSE)
}
