% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_cls_tab.R
\docType{methods}
\name{net_cls_tab,mina-method}
\alias{net_cls_tab,mina-method}
\title{Get the cluster table @cls_tab from quantitative table @tab and network
clustering results @cls.}
\usage{
\S4method{net_cls_tab}{mina}(x, uw = FALSE)
}
\arguments{
\item{x}{An object of class `mina` with @norm_tab and @cls defined.}

\item{uw}{By summing up the number of present components of each cluster
instead of relative abundance, default FALSE.}
}
\value{
x The same `mina` object with @cls_tab added.
}
\description{
Get the cluster table @cls_tab from quantitative table @tab and network
clustering results @cls.
}
\examples{
x <- net_cls_tab(x)
x <- net_cls_tab(x, uw = TRUE)
}
