% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{net_cls_tab}
\alias{net_cls_tab}
\title{Get the cluster table @cls_tab from @tab and @cls.}
\usage{
net_cls_tab(x, uw = FALSE)
}
\arguments{
\item{x}{An object of class `mina` with @tab and @cls defined.}

\item{uw}{By summing up the number of present components of each cluster
instead of relative abundance, default FALSE.}
}
\description{
Get the cluster table @cls_tab from @tab and @cls.
}
\examples{
x <- net_cls_tab(x)
x <- net_cls_tab(x, uw = TRUE)
}
