% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{net_cls_tab}
\alias{net_cls_tab}
\title{Get the cluster table @cls_tab from @norm and @cls.}
\usage{
net_cls_tab(x, uw = FALSE)
}
\arguments{
\item{x}{An object of class `mina` with @norm and @cls defined.}

\item{uw}{By summing up the number of present components of each cluster
instead of relative abundances, default is FALSE.}
}
\description{
Get the cluster table @cls_tab from @norm and @cls.
}
\examples{
data(maize)
maize@tab <- maize@tab[1 : 500, 1 : 200]
maize <- norm_tab(maize, method = "raref", depth = 100)
maize <- fit_tabs(maize)
maize <- adj(maize, method = "spearman")
maize <- net_cls(maize, method = "mcl", cutoff = 0.5)
maize <- net_cls_tab(maize)
maize <- net_cls(maize, method = "ap", cutoff = 0.5)
maize <- net_cls_tab(maize)
}
