% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_tab.R
\name{norm_by_raref}
\alias{norm_by_raref}
\title{Function for normalization by rarefying the samples into the same depth,
modified from \code{\link[phyloseq]{rarefy_even_depth}}.}
\usage{
norm_by_raref(x, depth = 1000, replace = TRUE)
}
\arguments{
\item{x}{A quantitative table with sample in columns and compostions in rows.}

\item{depth}{The depth for rarefying, 1000 by default.}

\item{replace}{Whether to sample with replacement (\code{TRUE}) or without
replacement (\code{FALSE}). Default \code{TRUE} for computational efficiency.}
}
\value{
A normalized quantitative table.
}
\description{
Function for normalization by rarefying the samples into the same depth,
modified from \code{\link[phyloseq]{rarefy_even_depth}}.
}
\keyword{internal}
