% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_tab.R
\docType{methods}
\name{norm_tab,matrix,character-method}
\alias{norm_tab,matrix,character-method}
\title{Normalize the quantitative table with matrix input.}
\usage{
\S4method{norm_tab}{matrix,character}(x, method, depth = 1000,
  replace = TRUE)
}
\arguments{
\item{x}{A matrix of the quantitative table.}

\item{method}{The method used for normalization.}

\item{depth}{The depth for subsampling by rarefying, using the minimum sample
depth by default.}

\item{replace}{Whether to sample with replacement (\code{TRUE} by default)
or without replacement (\code{FALSE}) when using method `raref`.}
}
\value{
norm_x Normalized matrix of the quantitative table.
}
\description{
Normalize the quantitative table with matrix input.
}
\examples{
norm_tab(x, method = "raref", depth = 1000, replace = TRUE)
norm_tab(x, method = "total")
}
