% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{norm_tab}
\alias{norm_tab}
\title{Normalize the @tab and obtain @norm for later analysis.}
\usage{
norm_tab(x, method, depth = 1000, replace = TRUE)
}
\arguments{
\item{x}{The input mina object with quantitative tab / a matrix needed to be
normalized.}

\item{method}{The method used for the normalization of quantitative table.}
}
\description{
Normalize the @tab and obtain @norm for later analysis.
}
\examples{
x <- norm_tab(x, method = "total")
}
