% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{norm_tab}
\alias{norm_tab}
\title{Normalize the slot `tab` for later analysis.}
\usage{
norm_tab(x, method, depth = 1000, replace = TRUE, multi = FALSE)
}
\arguments{
\item{x}{The input mina object with quantitative tab / a matrix needed to be
normalized.}

\item{method}{The method used for the normalization of quantitative table.}

\item{depth}{The depth for subsampling by rarefying, 1000 by default.}

\item{replace}{Whether to sample with replacement (\code{TRUE} by default) or
without replacement (\code{FALSE}) when using method `raref`.}

\item{multi}{Rarefy the table for multiple times, FALSE by default, indicate
the times of rarefaction want to be repeated, only validate for rarefaction.}
}
\value{
Normalized quantitative table.
}
\description{
Normalize the slot `tab` for later analysis.
}
\examples{
maize <- new("mina", tab = maize_asv2, des = maize_des2)
maize <- norm_tab(maize, method = "total")
maize <- norm_tab(maize, method = "raref")
maize <- norm_tab(maize, method = "raref", depth = 1000, replace = TRUE)
maize <- norm_tab(maize, method = "raref", depth = 1000, replace = TRUE,
multi = 99)
}
