% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_plot.R
\name{pcoa_plot}
\alias{pcoa_plot}
\title{Visulization of components distance / dissimilarity in k dimension.}
\usage{
pcoa_plot(x, des, match, color, shape = NULL)
}
\arguments{
\item{x}{An list generated by `dmr`.}

\item{des}{The corresponding descriptive table.}

\item{match}{The column name of the components IDs in `des` with exactly the
same as rownames in x.}

\item{color}{The column name in `des` to be used for different color groups.}

\item{shape}{The column name in `des` to be used for different shape groups,
default `NULL`.}
}
\value{
p The plotted figure.
}
\description{
Visulization of components distance / dissimilarity in k dimension.
}
\examples{
\dontrun{
data(maize)
maize <- norm_tab(maize, method = "raref")
maize <- fit_tabs(maize)
maize <- com_dis(maize, method = "bray")
maize <- dmr(maize)
asv_dmr <- maize@dmr
des <- maize@des
p1 <- com_plot(asv_dmr, des, match = "Sample_ID", color = "Compartment")
p2 <- com_plot(asv_dmr, des, match = "Sample_ID", color = "Host_genotype")
p3 <- com_plot(asv_dmr, des, match = "Sample_ID", color = "Compartment",
shape = "Soil")
}
}
