% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_tab.R
\name{rarefaction_subsample}
\alias{rarefaction_subsample}
\title{Rarefaction subsample function, one sample, modified from
\code{\link[phyloseq]{rarefaction_subsample}} which is a internal function.
Resample with replacement for default. As mentioned in
\code{\link[phyloseq]}, this is set for computational efficiency.}
\usage{
rarefaction_subsample(x, depth, replace = TRUE)
}
\arguments{
\item{x}{A column of quantitative table.}

\item{depth}{The depth for rarefying.}

\item{replace}{Whether to sample with or without replacement.}
}
\description{
Rarefaction subsample function, one sample, modified from
\code{\link[phyloseq]{rarefaction_subsample}} which is a internal function.
Resample with replacement for default. As mentioned in
\code{\link[phyloseq]}, this is set for computational efficiency.
}
\examples{
rarefaction_subsample(otu_table[, 1], depth = 1000, replace = FALSE)
}
\keyword{internal}
