% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_tab.R
\name{rarefaction_subsample}
\alias{rarefaction_subsample}
\title{Rarefaction subsample function, one sample, modified from
\code{\link[phyloseq]{rarefaction_subsample}} which is a internal function.
Resample with replacement for default. As mentioned in
\pkg{phyloseq}, this is set for computational efficiency.}
\usage{
rarefaction_subsample(x, depth = 1000, replace = TRUE)
}
\arguments{
\item{x}{A column of quantitative table.}

\item{depth}{The depth for rarefying, 1000 by default.}

\item{replace}{Whether to sample with or without replacement, \code{TRUE} by
default.}
}
\description{
Rarefaction subsample function, one sample, modified from
\code{\link[phyloseq]{rarefaction_subsample}} which is a internal function.
Resample with replacement for default. As mentioned in
\pkg{phyloseq}, this is set for computational efficiency.
}
\keyword{internal}
