% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_cls.R
\name{re_format_AP}
\alias{re_format_AP}
\title{Convert APResult (apcluster output) to dataframe.}
\usage{
re_format_AP(x)
}
\arguments{
\item{x}{an {APResult} object from \pkg{apcluster}.}
}
\value{
y A data frame with columns `ID`, `Exemplar`, `Cluster` and
`Cluster_size`.
}
\description{
Modified from https://rdrr.io/github/jefferis/flycircuit/src/R/clustering.R
#sym-as.data.frame.APResult
}
\keyword{internal}
