% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_dis.R
\name{sim_par}
\alias{sim_par}
\title{Function for community similarity calculation used by `tina`, modified from
https://github.com/defleury/Schmidt_et_al_2016_community_similarity/blob/
master/functions.community_similarity.R}
\usage{
sim_par(x, y, sim_method = "w_ja", threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An quantitative matrix.}

\item{y}{The Cij matrix, which is correlation matrix of adjusted sparcc
matrix of x.}

\item{sim_method}{The method for similarity, "w_ja" and "uw_ja" are
available for weighted and unweighted Jaccard similarity respectively.}

\item{threads}{(optional) The number of threads used for parallel running,
80 by default.}

\item{nblocks}{(optional) The number of row / column for splitted sub-matrix,
400 by default.}
}
\value{
s The output similarity matrix.
}
\description{
Function for community similarity calculation used by `tina`, modified from
https://github.com/defleury/Schmidt_et_al_2016_community_similarity/blob/
master/functions.community_similarity.R
}
\examples{
s <- sim_par(x, y, sim_method = "w_ja", threads = 80, nblocks = 400)
}
\keyword{internal}
