% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\name{sparcc}
\alias{sparcc}
\title{Function for `sparcc` correlation calculation. Modified from
https://github.com/defleury/Schmidt_et_al_2016_community_similarity/blob/
master/functions.community_similarity.R}
\usage{
sparcc(x, threads = 80, nblocks = 400)
}
\arguments{
\item{x}{An matrix for correlation calculation.}

\item{threads}{(optional) The number of threads used for parallel running, 80
by default.}

\item{nblocks}{(optional) The number of row / column for splitted sub-matrix,
400 by default.}
}
\value{
y The adjacency matrix.
}
\description{
Function for `sparcc` correlation calculation. Modified from
https://github.com/defleury/Schmidt_et_al_2016_community_similarity/blob/
master/functions.community_similarity.R
}
\examples{
y <- sparcc(x, threads = 80, nblocks = 400)
}
\keyword{internal}
