% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_dis.R
\name{tina,matrix,character,character-method}
\alias{tina,matrix,character,character-method}
\title{Function for `tina` dissimilarity calculation. Modified from Schmidt et al.,
2016. Person and Spearman could be used for correlation and weighted and
unweighted Jaccard could be used for similarity calculation.}
\usage{
\S4method{tina}{matrix,character,character}(
  x,
  cor_method = "spearman",
  sim_method = "w_ja",
  threads = 80,
  nblocks = 400
)
}
\arguments{
\item{x}{A matrix for dissimilarity calculation.}

\item{cor_method}{The method for correlation, "pearson" and "spearman" are
available.}

\item{sim_method}{The method for similarity, "w_ja" and "uw_ja" are
available for weighted and unweighted Jaccard similarity respectively.}

\item{threads}{The number of threads used for parallel running, 80 by
default.}

\item{nblocks}{The number of row and column for splitted sub-matrix, 400 by
default.}
}
\value{
t The output `tina` dissimilarity matrix.
}
\description{
Function for `tina` dissimilarity calculation. Modified from Schmidt et al.,
2016. Person and Spearman could be used for correlation and weighted and
unweighted Jaccard could be used for similarity calculation.
}
\examples{
\dontrun{
data(maize)
maize@tab <- maize@tab[1 : 1000, 1 : 200]
maize <- norm_tab(maize, method = "raref", depth = 100)
maize <- fit_tabs(maize)
asv_norm <- maize@norm
asv_norm[is.na(asv_norm)] <- 0
asv_tina <- tina(asv_norm, cor_method = "spearman", sim_method = "w_ja",
threads = 8, nblocks = 40)
}
}
