% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_dis.R
\name{tina}
\alias{tina}
\title{Function for `tina` dissimilarity / distance calculation. Modified from
https://github.com/defleury/Schmidt_et_al_2016_community_similarity/blob/
master/functions.community_similarity.R
Pearson / Spearman could be used for calculating correlation and weighted /
unweighted Jaccard could be used for the calculation of similarity.}
\usage{
tina(x, cor_method = "spearman", sim_method = "w_ja", threads = 80,
  nblocks = 400)
}
\arguments{
\item{x}{An matrix for `tina` dissimilarity calculation.}

\item{cor_method}{The method for correlation, "pearson" and "spearman" are
available.}

\item{sim_method}{The method for similarity, "w_ja" and "uw_ja" are
available for weighted and unweighted Jaccard similarity respectively.}

\item{threads}{(optional) The number of threads used for parallel running,
80 by default.}

\item{nblocks}{(optional) The number of row / column for splitted sub-matrix,
400 by default.}
}
\value{
t The output `tina` dissimilarity matrix.
}
\description{
Function for `tina` dissimilarity / distance calculation. Modified from
https://github.com/defleury/Schmidt_et_al_2016_community_similarity/blob/
master/functions.community_similarity.R
Pearson / Spearman could be used for calculating correlation and weighted /
unweighted Jaccard could be used for the calculation of similarity.
}
\examples{
\dontrun{
data(maize)
maize <- norm_tab(maize, method = "raref")
x <- maize@norm[1:500, 1:300]
t <- tina(x, cor_method = "spearman", sim_method = "w_ja", threads = 8,
          nblocks = 40)
}
}
